requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
pydantic
python-dotenv
pyyaml
pytest
pytest-asyncio
respx